Disasters no longer spawn on same tile

+++ Disaster.h
@@ -37,7 +37,7 @@ public:
 
 protected:
     int m_difficultyChance; // Chance of spreading depending on selected difficulty 
-    int m_health;
+    float m_health;
     bool m_spawned = false;
     sf::Sprite m_disasterSprite;
     sf::Texture m_spriteTexture;
